--- a/transport/grpc/response_writer_test.go
+++ b/transport/grpc/response_writer_test.go
@@ -6,7 +6,82 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/yarpc/api/transport"
+	"google.golang.org/grpc/metadata"
 )
+
+func Test_AddHeader(t *testing.T) {
+	tests := []struct {
+		name   string
+		key    string
+		value  string
+		start  metadata.MD
+		result metadata.MD
+		err    bool
+	}{
+		{
+			name:   "empty key",
+			key:    "foo",
+			result: metadata.MD{},
+		},
+		{
+			name:   "lowercase",
+			key:    "foo",
+			value:  "bar",
+			result: metadata.New(map[string]string{"foo": "bar"}),
+		},
+		{
+			name:   "titlecase",
+			key:    "Foo",
+			value:  "Bar",
+			result: metadata.New(map[string]string{"foo": "Bar"}),
+		},
+		{
+			name:   "restricted key",
+			key:    "rpc-foo",
+			result: metadata.MD{},
+			err:    true,
+		},
+		{
+			name:  "duplicate key",
+			key:   "foo",
+			value: "bar",
+			start: metadata.New(map[string]string{"foo": "bar"}),
+			err:   true,
+		},
+	}
+
+	t.Run("AddHeader", func(t *testing.T) {
+		for _, test := range tests {
+			t.Run(test.name, func(t *testing.T) {
+				rw := newResponseWriter()
+				rw.md = test.start
+				rw.AddHeader(test.key, test.value)
+				if test.err {
+					assert.Error(t, rw.headerErr)
+				} else {
+					assert.NoError(t, rw.headerErr)
+					assert.Equal(t, test.result, rw.md)
+				}
+			})
+		}
+	})
+
+	t.Run("AddHeaders", func(t *testing.T) {
+		for _, test := range tests {
+			t.Run(test.name, func(t *testing.T) {
+				rw := newResponseWriter()
+				rw.md = test.start
+				rw.AddHeaders(transport.HeadersFromMap(map[string]string{test.key: test.value}))
+				if test.err {
+					assert.Error(t, rw.headerErr)
+				} else {
+					assert.NoError(t, rw.headerErr)
+					assert.Equal(t, test.result, rw.md)
+				}
+			})
+		}
+	})
+}
 
 func Benchmark_ResponseWriter_AddHeaders(b *testing.B) {
 	b.Run("lowercase", func(b *testing.B) {
@@ -48,18 +123,6 @@ func Benchmark_ResponseWriter_AddHeaders(b *testing.B) {
 		}
 		assert.NoError(b, rw.headerErr)
 	})
-}
-
-func Test_addHeaders(t *testing.T) {
-	// t.Run("No ")
-	rw := newResponseWriter()
-	for i := 0; i < 10; i++ {
-		rw.AddHeaders(transport.NewHeadersWithCapacity(1).With(
-			"foo", "bar",
-		))
-	}
-	assert.NoError(t, rw.headerErr)
-
 }
 
 func Benchmark_ResponseWriter_AddHeader(b *testing.B) {