--- a/transport/http/handler_test.go
+++ b/transport/http/handler_test.go
@@ -454,6 +454,46 @@ func TestResponseWriter(t *testing.T) {
 	assert.Equal(t, strconv.Itoa(int(appErrCode)), recorder.Header().Get(_applicationErrorCodeHeader))
 }
 
+func Benchmark_ResponseWriter_AddHeaders(b *testing.B) {
+	b.Run("lowercase", func(b *testing.B) {
+		recorder := httptest.NewRecorder()
+		rw := newResponseWriter(recorder)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeaders(transport.NewHeadersWithCapacity(1).With(
+				"abc", "",
+			))
+		}
+	})
+
+	b.Run("titlecase", func(b *testing.B) {
+		recorder := httptest.NewRecorder()
+		rw := newResponseWriter(recorder)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeaders(transport.NewHeadersWithCapacity(1).With(
+				"Abc", "",
+			))
+		}
+	})
+}
+
+func Benchmark_ResponseWriter_AddHeader(b *testing.B) {
+	b.Run("lowercase", func(b *testing.B) {
+		recorder := httptest.NewRecorder()
+		rw := newResponseWriter(recorder)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeader("abc", "")
+		}
+	})
+
+	b.Run("titlecase", func(b *testing.B) {
+		recorder := httptest.NewRecorder()
+		rw := newResponseWriter(recorder)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeader("Abc", "")
+		}
+	})
+}
+
 func TestTruncatedHeader(t *testing.T) {
 	tests := []struct {
 		name         string