--- a/transport/tchannel/handler_test.go
+++ b/transport/tchannel/handler_test.go
@@ -648,6 +648,42 @@ func TestResponseWriterEmptyBodyHeaders(t *testing.T) {
 	assert.False(t, res.applicationError, "application error must be false")
 }
 
+func Benchmark_ResponseWriter_AddHeaders(b *testing.B) {
+	b.Run("lowercase", func(b *testing.B) {
+		rw := newHandlerWriter(nil, tchannel.Raw, canonicalizedHeaderCase)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeaders(transport.NewHeadersWithCapacity(1).With(
+				"abc", "",
+			))
+		}
+	})
+
+	b.Run("titlecase", func(b *testing.B) {
+		rw := newHandlerWriter(nil, tchannel.Raw, canonicalizedHeaderCase)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeaders(transport.NewHeadersWithCapacity(1).With(
+				"Abc", "",
+			))
+		}
+	})
+}
+
+func Benchmark_ResponseWriter_AddHeader(b *testing.B) {
+	b.Run("lowercase", func(b *testing.B) {
+		rw := newHandlerWriter(nil, tchannel.Raw, canonicalizedHeaderCase)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeader("abc", "")
+		}
+	})
+
+	b.Run("titlecase", func(b *testing.B) {
+		rw := newHandlerWriter(nil, tchannel.Raw, canonicalizedHeaderCase)
+		for i := 0; i < b.N; i++ {
+			rw.AddHeader("Abc", "")
+		}
+	})
+}
+
 func TestGetSystemError(t *testing.T) {
 	tests := []struct {
 		giveErr  error