@rabbbit 2022-02-28
I know very little about yarpc, but - is the `AddHeader` a viable option?

Looking at https://github.com/yarpc/yarpc-go/blob/dev/api/transport/response.go#L55, we would need to modify the interface, which is a breaking change? 

Sorry, I didn't call it out, I didn't realize the impact of it earlier.