@gopalchouhan 2022-03-21
I agree with @rabbbit responseWriter implements the interface **transport.ResponseWriter**, and  transport.ResponseWrite interface has to be modified to make it work. Modifying an existing common interface could be a breaking change and is something needs to be planned very carefully while keeping all users in mind that include internal as well as external users.
Also we should evaluate the value this change brings against the sensitivity of this change.