@jstroem 2022-03-21
I agree, the contribution of this PR is not too much and since it requires changing the interface it is a big thing.

I still think it makes sense to do the optimisations on picking the right header-key so you don't have to transport strings unnecessarily - so I'll open a new PR for that.

Closing this one.